Revamp the StatefulCommandHandlingModule to use an EntityBuilder
Revamp the StatefulCommandHandlingModule to use an EntityBuilder.
Although we lose the "flat fluent API" for registering entities like
this, it makes it straightforward to provide other entity builders. For
example, an event sourced build. This eliminates **a lot** of duplicate
code that would otherwise be required to register event sourced modules,
 since we can reuse the StatefulCommandHandlingModule entirely for event
  sourcing applications.

#3345
